PART 1 AND PART 2/src/part/pkg1/and/part/pkg2/UserAuthentication.java
/*
 * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
 * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
 */
package part.pkg1.and.part.pkg2;

/**
 *
 * @author RC_Student_lab
 */
public class UserAuthentication {
private String username;
    private String password;
    private String firstName;
    private String lastName;

    public UserAuthentication(String username, String password, String firstName, String lastName) {
        this.username = username;
        this.password = password;
        this.firstName = firstName;
        this.lastName = lastName;
    }

    // Method to check if username is correctly formatted
    public boolean checkUserName(String username) {
        return username.contains("_") && username.length() <= 5;
    }

    // Method to check if password meets complexity requirements
    public boolean checkPasswordComplexity(String password) {
        return password.length() >= 8 &&
               password.matches(".*[A-Z].*") &&
               password.matches(".*[0-9].*") &&
               password.matches(".*[@#$%^&+=].*");
    }

    // Method to register user
    public String registerUser(String username, String password) {
        if (!checkUserName(username)) {
            return "Username is not correctly formatted, please ensure that your username contains an underscore and is no more than 5 characters in length.";
        }
        if (!checkPasswordComplexity(password)) {
            return "Password is not correctly formatted, please ensure that the password contains at least 8 characters, a capital letter, a number, and a special character.";
        }
        // Registration successful
        return "User registered successfully!";
    }

    // Method to login
    public boolean login(String enteredUsername, String enteredPassword) {
        return enteredUsername.equals(this.username) && enteredPassword.equals(this.password);
    }

    // Getters for first name and last name
    public String getFirstName() {
        return firstName;
    }

    public String getLastName() {
        return lastName;
    }
}